Refactor video output
