+++ v4l2out.cpp
@@ -64,296 +64,272 @@ public:
 	}
 };
 
-class Video_out
+Video_out::Video_out(const char *devNameIn)
 {
-public:
-	Video_out_manager *self;
-	std::string devName;
-	int stop;
-	int stopped;
-	pthread_mutex_t lock;
-	int verbose;
-	std::vector<class SendFrameArgs> sendFrameArgs;
-	std::vector<const char *> sendFrameBuffer;
-	struct timespec lastFrameTime;
-	int fdwr;
-	int framesize;
-	unsigned char *currentFrame;
-	int outputWidth;
-	int outputHeight;
-	std::string outputPxFmt;
+	this->fdwr = 0;
+	framesize = 0;
+	stop = 0;
+	stopped = 1;
+	verbose = 1;
+	this->devName = devNameIn;
+	pthread_mutex_init(&lock, NULL);
+	currentFrame = NULL;
+	outputWidth = 640;
+	outputHeight = 480;
+	outputPxFmt = "YUYV";
 
-	Video_out(const char *devNameIn)
-	{
-		this->fdwr = 0;
-		framesize = 0;
-		stop = 0;
-		stopped = 1;
-		verbose = 1;
-		this->devName = devNameIn;
-		pthread_mutex_init(&lock, NULL);
-		currentFrame = NULL;
-		outputWidth = 640;
-		outputHeight = 480;
-		outputPxFmt = "YUYV";
+	clock_gettime(CLOCK_MONOTONIC, &lastFrameTime);
 
-		clock_gettime(CLOCK_MONOTONIC, &lastFrameTime);
-
-		struct sigevent sevp;
-		memset(&sevp, 0, sizeof(struct sigevent));
-		sevp.sigev_notify = SIGEV_NONE;
+	struct sigevent sevp;
+	memset(&sevp, 0, sizeof(struct sigevent));
+	sevp.sigev_notify = SIGEV_NONE;
 	
+}
+
+Video_out::~Video_out()
+{
+	for(unsigned i=0; i<this->sendFrameBuffer.size(); i++)
+	{
+		delete [] this->sendFrameBuffer[i];
+	}
+	this->sendFrameBuffer.clear();
+
+	if(this->currentFrame!=NULL)
+		delete [] this->currentFrame;
+	this->currentFrame = NULL;
+
+	pthread_mutex_destroy(&lock);
+}
+
+void Video_out::SendFrameInternal()
+{
+	const char* buff = NULL;
+	class SendFrameArgs args;
+			
+	pthread_mutex_lock(&this->lock);
+	if(this->sendFrameBuffer.size()>=1)
+	{
+		//Get oldest frame
+		buff = this->sendFrameBuffer[0];
+		args = this->sendFrameArgs[0];
+
+		//Remove frame from buffer
+		this->sendFrameBuffer.erase(this->sendFrameBuffer.begin());
+		this->sendFrameArgs.erase(this->sendFrameArgs.begin());
+	}
+	pthread_mutex_unlock(&this->lock);
+
+	//Check time since previous frame send
+	struct timespec tp;
+	clock_gettime(CLOCK_MONOTONIC, &tp);
+	long int secSinceLastFrame = tp.tv_sec - this->lastFrameTime.tv_sec;
+	long int nsecSinceLastFrame = tp.tv_nsec - this->lastFrameTime.tv_nsec;
+	if(nsecSinceLastFrame < 0)
+	{
+		secSinceLastFrame -= 1;
+		nsecSinceLastFrame *= -1;
 	}
 
-	virtual ~Video_out()
+	if(buff != NULL)
 	{
-		for(unsigned i=0; i<this->sendFrameBuffer.size(); i++)
+		//Convert new frame to correct size and pixel format
+		assert(strcmp(args.pxFmt.c_str(), "RGB24")==0);
+		unsigned resizeBuffLen = this->outputWidth * this->outputHeight * 3;
+		char *buffResize = new char[resizeBuffLen];
+		memset(buffResize, 0, resizeBuffLen);
+		for(unsigned x = 0; x < this->outputWidth; x++)
 		{
-			delete [] this->sendFrameBuffer[i];
+			if (x >= args.width) continue;
+			for(unsigned y = 0; y < this->outputHeight; y++)
+			{
+				if (y >= args.height) continue;
+				buffResize[y * this->outputWidth * 3 + x * 3] = buff[y * args.width * 3 + x * 3];
+				buffResize[y * this->outputWidth * 3 + x * 3 + 1] = buff[y * args.width * 3 + x * 3 + 1];
+				buffResize[y * this->outputWidth * 3 + x * 3 + 2] = buff[y * args.width * 3 + x * 3 + 2];
+			}
 		}
-		this->sendFrameBuffer.clear();
 
+		unsigned char *buffOut = NULL;
+		unsigned buffOutLen = 0;
+		DecodeFrame((unsigned char *)buffResize, resizeBuffLen, 
+			args.pxFmt.c_str(),
+			this->outputWidth, this->outputHeight,
+			this->outputPxFmt.c_str(),
+			&buffOut,
+			&buffOutLen);
+
+		assert(buffOutLen == this->framesize);
+
+		//Replace current frame with new encoded frame
 		if(this->currentFrame!=NULL)
 			delete [] this->currentFrame;
-		this->currentFrame = NULL;
+		this->currentFrame = buffOut;
 
-		pthread_mutex_destroy(&lock);
+		delete [] buffResize;
+
 	}
 
-protected:
-
-	void SendFrameInternal()
+	//If we have no data, initialise with a blank frame
+	if(this->currentFrame==NULL)
 	{
-		const char* buff = NULL;
-		class SendFrameArgs args;
-			
-		pthread_mutex_lock(&this->lock);
-		if(this->sendFrameBuffer.size()>=1)
-		{
-			//Get oldest frame
-			buff = this->sendFrameBuffer[0];
-			args = this->sendFrameArgs[0];
-
-			//Remove frame from buffer
-			this->sendFrameBuffer.erase(this->sendFrameBuffer.begin());
-			this->sendFrameArgs.erase(this->sendFrameArgs.begin());
-		}
-		pthread_mutex_unlock(&this->lock);
-
-		//Check time since previous frame send
-		struct timespec tp;
-		clock_gettime(CLOCK_MONOTONIC, &tp);
-		long int secSinceLastFrame = tp.tv_sec - this->lastFrameTime.tv_sec;
-		long int nsecSinceLastFrame = tp.tv_nsec - this->lastFrameTime.tv_nsec;
-		if(nsecSinceLastFrame < 0)
-		{
-			secSinceLastFrame -= 1;
-			nsecSinceLastFrame *= -1;
-		}
-
-		if(buff != NULL)
-		{
-			//Convert new frame to correct size and pixel format
-			assert(strcmp(args.pxFmt.c_str(), "RGB24")==0);
-			unsigned resizeBuffLen = this->outputWidth * this->outputHeight * 3;
-			char *buffResize = new char[resizeBuffLen];
-			memset(buffResize, 0, resizeBuffLen);
-			for(unsigned x = 0; x < this->outputWidth; x++)
-			{
-				if (x >= args.width) continue;
-				for(unsigned y = 0; y < this->outputHeight; y++)
-				{
-					if (y >= args.height) continue;
-					buffResize[y * this->outputWidth * 3 + x * 3] = buff[y * args.width * 3 + x * 3];
-					buffResize[y * this->outputWidth * 3 + x * 3 + 1] = buff[y * args.width * 3 + x * 3 + 1];
-					buffResize[y * this->outputWidth * 3 + x * 3 + 2] = buff[y * args.width * 3 + x * 3 + 2];
-				}
-			}
-
-			unsigned char *buffOut = NULL;
-			unsigned buffOutLen = 0;
-			DecodeFrame((unsigned char *)buffResize, resizeBuffLen, 
-				args.pxFmt.c_str(),
-				this->outputWidth, this->outputHeight,
-				this->outputPxFmt.c_str(),
-				&buffOut,
-				&buffOutLen);
-
-			assert(buffOutLen == this->framesize);
-
-			//Replace current frame with new encoded frame
-			if(this->currentFrame!=NULL)
-				delete [] this->currentFrame;
-			this->currentFrame = buffOut;
-
-			delete [] buffResize;
-
-		}
-
-		//If we have no data, initialise with a blank frame
-		if(this->currentFrame==NULL)
-		{
-			this->currentFrame = new unsigned char[this->framesize];
-			memset(this->currentFrame, 0, this->framesize);
-		}
-
-		int timeElapsed = secSinceLastFrame>=1;
-
-		if(timeElapsed || buff != NULL)
-		{
-			//Send frame update due to time elapse
-			if(timeElapsed)
-				printf("Write frame due to elapse time\n");
-			write(this->fdwr, this->currentFrame, this->framesize);
-
-			this->lastFrameTime = tp;
-		}
-
-		//Free image buffer
-		if(buff!=NULL)
-			delete [] buff;
+		this->currentFrame = new unsigned char[this->framesize];
+		memset(this->currentFrame, 0, this->framesize);
 	}
 
-public:
-	void Run()
+	int timeElapsed = secSinceLastFrame>=1;
+
+	if(timeElapsed || buff != NULL)
 	{
-		if(verbose) printf("Thread started: %s\n", this->devName.c_str());
-		int running = 1;
-		pthread_mutex_lock(&this->lock);
-		this->stopped = 0;
-		pthread_mutex_unlock(&this->lock);
+		//Send frame update due to time elapse
+		if(timeElapsed)
+			printf("Write frame due to elapse time\n");
+		write(this->fdwr, this->currentFrame, this->framesize);
 
-		this->fdwr = open(this->devName.c_str(), O_RDWR);
-		assert(fdwr >= 0);
+		this->lastFrameTime = tp;
+	}
 
-		struct v4l2_capability vid_caps;
-		int ret_code = ioctl(this->fdwr, VIDIOC_QUERYCAP, &vid_caps);
-		assert(ret_code != -1);
+	//Free image buffer
+	if(buff!=NULL)
+		delete [] buff;
+}
+
+void Video_out::Run()
+{
+	if(verbose) printf("Thread started: %s\n", this->devName.c_str());
+	int running = 1;
+	pthread_mutex_lock(&this->lock);
+	this->stopped = 0;
+	pthread_mutex_unlock(&this->lock);
+
+	this->fdwr = open(this->devName.c_str(), O_RDWR);
+	assert(fdwr >= 0);
+
+	struct v4l2_capability vid_caps;
+	int ret_code = ioctl(this->fdwr, VIDIOC_QUERYCAP, &vid_caps);
+	assert(ret_code != -1);
 	
-		struct v4l2_format vid_format;
-		memset(&vid_format, 0, sizeof(vid_format));
+	struct v4l2_format vid_format;
+	memset(&vid_format, 0, sizeof(vid_format));
 
-		ret_code = ioctl(this->fdwr, VIDIOC_G_FMT, &vid_format);
-		if(verbose)print_format(&vid_format);
+	ret_code = ioctl(this->fdwr, VIDIOC_G_FMT, &vid_format);
+	if(verbose)print_format(&vid_format);
 
-		int lw = 0;
-		int fw = 0;
-		if(strcmp(this->outputPxFmt.c_str(), "YVU420")==0)
-		{
-			lw = this->outputWidth; /* ??? */
-			fw = ROUND_UP_4 (this->outputWidth) * ROUND_UP_2 (this->outputHeight);
-			fw += 2 * ((ROUND_UP_8 (this->outputWidth) / 2) * (ROUND_UP_2 (this->outputHeight) / 2));
-		}
+	int lw = 0;
+	int fw = 0;
+	if(strcmp(this->outputPxFmt.c_str(), "YVU420")==0)
+	{
+		lw = this->outputWidth; /* ??? */
+		fw = ROUND_UP_4 (this->outputWidth) * ROUND_UP_2 (this->outputHeight);
+		fw += 2 * ((ROUND_UP_8 (this->outputWidth) / 2) * (ROUND_UP_2 (this->outputHeight) / 2));
+	}
 
-		if(strcmp(this->outputPxFmt.c_str(), "YUYV")==0 
-			|| strcmp(this->outputPxFmt.c_str(), "UYVY")==0 )
-		{
-			lw = (ROUND_UP_2 (this->outputWidth) * 2);
-			fw = lw * this->outputHeight;
-		}
+	if(strcmp(this->outputPxFmt.c_str(), "YUYV")==0 
+		|| strcmp(this->outputPxFmt.c_str(), "UYVY")==0 )
+	{
+		lw = (ROUND_UP_2 (this->outputWidth) * 2);
+		fw = lw * this->outputHeight;
+	}
 
-		vid_format.type = V4L2_BUF_TYPE_VIDEO_OUTPUT;
-		vid_format.fmt.pix.width = this->outputWidth;
-		vid_format.fmt.pix.height = this->outputHeight;
-		vid_format.fmt.pix.pixelformat = 0;
-		if(strcmp(this->outputPxFmt.c_str(), "YUYV")==0)
-			vid_format.fmt.pix.pixelformat = V4L2_PIX_FMT_YUYV;
-		if(strcmp(this->outputPxFmt.c_str(), "UYVY")==0)
-			vid_format.fmt.pix.pixelformat = V4L2_PIX_FMT_UYVY;
-		if(strcmp(this->outputPxFmt.c_str(), "YVU420")==0)
-			vid_format.fmt.pix.pixelformat = V4L2_PIX_FMT_YVU420;
-		if(strcmp(this->outputPxFmt.c_str(), "RGB24")==0)
-			vid_format.fmt.pix.pixelformat = V4L2_PIX_FMT_RGB24;
+	vid_format.type = V4L2_BUF_TYPE_VIDEO_OUTPUT;
+	vid_format.fmt.pix.width = this->outputWidth;
+	vid_format.fmt.pix.height = this->outputHeight;
+	vid_format.fmt.pix.pixelformat = 0;
+	if(strcmp(this->outputPxFmt.c_str(), "YUYV")==0)
+		vid_format.fmt.pix.pixelformat = V4L2_PIX_FMT_YUYV;
+	if(strcmp(this->outputPxFmt.c_str(), "UYVY")==0)
+		vid_format.fmt.pix.pixelformat = V4L2_PIX_FMT_UYVY;
+	if(strcmp(this->outputPxFmt.c_str(), "YVU420")==0)
+		vid_format.fmt.pix.pixelformat = V4L2_PIX_FMT_YVU420;
+	if(strcmp(this->outputPxFmt.c_str(), "RGB24")==0)
+		vid_format.fmt.pix.pixelformat = V4L2_PIX_FMT_RGB24;
 
-		vid_format.fmt.pix.sizeimage = lw;
-		vid_format.fmt.pix.field = V4L2_FIELD_NONE;
-		vid_format.fmt.pix.bytesperline = fw;
-		vid_format.fmt.pix.colorspace = V4L2_COLORSPACE_SRGB;
+	vid_format.fmt.pix.sizeimage = lw;
+	vid_format.fmt.pix.field = V4L2_FIELD_NONE;
+	vid_format.fmt.pix.bytesperline = fw;
+	vid_format.fmt.pix.colorspace = V4L2_COLORSPACE_SRGB;
 
-		if(verbose)print_format(&vid_format);
+	if(verbose)print_format(&vid_format);
 
-		ret_code = ioctl(this->fdwr, VIDIOC_S_FMT, &vid_format);
+	ret_code = ioctl(this->fdwr, VIDIOC_S_FMT, &vid_format);
 
-		assert(ret_code != -1);
+	assert(ret_code != -1);
 
-		this->framesize = vid_format.fmt.pix.sizeimage;
-		int linewidth = vid_format.fmt.pix.bytesperline;
-		if(verbose)printf("frame: format=%s\tsize=%d\n", this->outputPxFmt.c_str(), framesize);
+	this->framesize = vid_format.fmt.pix.sizeimage;
+	int linewidth = vid_format.fmt.pix.bytesperline;
+	if(verbose)printf("frame: format=%s\tsize=%d\n", this->outputPxFmt.c_str(), framesize);
 
+	try
+	{
+	while(running)
+	{
+		usleep(1000);
+
+		this->SendFrameInternal();
+
+		pthread_mutex_lock(&this->lock);
 		try
 		{
-		while(running)
-		{
-			usleep(1000);
-
-			this->SendFrameInternal();
-
-			pthread_mutex_lock(&this->lock);
-			try
-			{
-
-				running = !this->stop;
-			}
-			catch(std::exception &err)
-			{
-				if(verbose) printf("An exception has occured: %s\n", err.what());
-				running = 0;
-			}
-			pthread_mutex_unlock(&this->lock);
-		}
+			running = !this->stop;
 		}
 		catch(std::exception &err)
 		{
 			if(verbose) printf("An exception has occured: %s\n", err.what());
+			running = 0;
 		}
-
-		if(verbose) printf("Thread stopping\n");
-		pthread_mutex_lock(&this->lock);
-		this->stopped = 1;
 		pthread_mutex_unlock(&this->lock);
 	}
+	}
+	catch(std::exception &err)
+	{
+		if(verbose) printf("An exception has occured: %s\n", err.what());
+	}
 
-	void SendFrame(const char *imgIn, unsigned imgLen, const char *pxFmt, int width, int height)
+	if(verbose) printf("Thread stopping\n");
+	pthread_mutex_lock(&this->lock);
+	this->stopped = 1;
+	pthread_mutex_unlock(&this->lock);
+}
+
+void Video_out::SendFrame(const char *imgIn, unsigned imgLen, const char *pxFmt, int width, int height)
+{
+	pthread_mutex_lock(&this->lock);
+	if(verbose) printf("SendFrame %i %s %i %i\n", imgLen, pxFmt, width, height);
+
+	//Take a shallow copy of the buffer and keep for worker thread
+	char *buffCpy = new char[imgLen];
+	memcpy(buffCpy, imgIn, imgLen);
+	this->sendFrameBuffer.push_back(buffCpy);
+
+	class SendFrameArgs sendFrameArgsTmp;
+	sendFrameArgsTmp.imgLen = imgLen;
+	sendFrameArgsTmp.pxFmt = pxFmt;
+	sendFrameArgsTmp.width = width;
+	sendFrameArgsTmp.height = height;
+	this->sendFrameArgs.push_back(sendFrameArgsTmp);
+
+	pthread_mutex_unlock(&this->lock);
+}
+
+void Video_out::Stop()
+{
+	pthread_mutex_lock(&this->lock);
+	this->stop = 1;
+	pthread_mutex_unlock(&this->lock);
+}
+
+int Video_out::WaitForStop()
+{
+	this->Stop();
+	while(1)
 	{
 		pthread_mutex_lock(&this->lock);
-		if(verbose) printf("SendFrame %i %s %i %i\n", imgLen, pxFmt, width, height);
-
-		//Take a shallow copy of the buffer and keep for worker thread
-		char *buffCpy = new char[imgLen];
-		memcpy(buffCpy, imgIn, imgLen);
-		this->sendFrameBuffer.push_back(buffCpy);
-
-		class SendFrameArgs sendFrameArgsTmp;
-		sendFrameArgsTmp.imgLen = imgLen;
-		sendFrameArgsTmp.pxFmt = pxFmt;
-		sendFrameArgsTmp.width = width;
-		sendFrameArgsTmp.height = height;
-		this->sendFrameArgs.push_back(sendFrameArgsTmp);
-
+		int s = this->stopped;
 		pthread_mutex_unlock(&this->lock);
-	}
 
-	void Stop()
-	{
-		pthread_mutex_lock(&this->lock);
-		this->stop = 1;
-		pthread_mutex_unlock(&this->lock);
+		if(s) return 1;
+		usleep(10000);
 	}
-
-	int WaitForStop()
-	{
-		this->Stop();
-		while(1)
-		{
-			pthread_mutex_lock(&this->lock);
-			int s = this->stopped;
-			pthread_mutex_unlock(&this->lock);
-
-			if(s) return 1;
-			usleep(10000);
-		}
-	}
-};
+}
 
 void *Video_out_manager_Worker_thread(void *arg)
 {
@@ -365,145 +341,3 @@ void *Video_out_manager_Worker_thread(void *arg)
 
 // *****************************************************************
 
-int Video_out_manager_init(Video_out_manager *self, PyObject *args,
-		PyObject *kwargs)
-{
-	self->threads = new std::map<std::string, class Video_out *>;
-	return 0;
-}
-
-void Video_out_manager_dealloc(Video_out_manager *self)
-{
-	//Stop high level threads
-	for(std::map<std::string, class Video_out *>::iterator it = self->threads->begin(); 
-		it != self->threads->end(); it++)
-	{
-		it->second->Stop();
-		it->second->WaitForStop();
-	}
-
-	delete self->threads;
-	self->threads = NULL;
-	self->ob_type->tp_free((PyObject *)self);
-}
-
-PyObject *Video_out_manager_open(Video_out_manager *self, PyObject *args)
-{
-	//Process arguments
-	const char *devarg = NULL;
-	const char *pxFmtIn = NULL;
-	int widthIn = 0;
-	int heightIn = 0;
-
-	if(!PyArg_ParseTuple(args, "ssii", &devarg, &pxFmtIn, &widthIn, &heightIn))
-	{
-		PyErr_Format(PyExc_RuntimeError, "Incorrect arguments to function.");
-		Py_RETURN_NONE;
-	}
-
-	//Create worker thread
-	pthread_t thread;
-	Video_out *threadArgs = new Video_out(devarg);
-	(*self->threads)[devarg] = threadArgs;
-	threadArgs->self = self;
-	threadArgs->outputWidth = widthIn;
-	threadArgs->outputHeight = heightIn;
-	threadArgs->outputPxFmt = pxFmtIn;
-
-	pthread_create(&thread, NULL, Video_out_manager_Worker_thread, threadArgs);
-
-	Py_RETURN_NONE;
-}
-
-PyObject *Video_out_manager_Send_frame(Video_out_manager *self, PyObject *args)
-{
-	//printf("Video_out_manager_Send_frame\n");
-	//dev = '\\dev\\video0', img, pixel_format, width, height
-
-	//Process arguments
-	const char *devarg = NULL;
-	const char *imgIn = NULL;
-	const char *pxFmtIn = NULL;
-	int widthIn = 0;
-	int heightIn = 0;
-
-	if(PyObject_Length(args) < 5)
-	{
-		PyErr_Format(PyExc_RuntimeError, "Too few arguments.");
-		Py_RETURN_NONE;
-	}
-
-	PyObject *pydev = PyTuple_GetItem(args, 0);
-	devarg = PyString_AsString(pydev);
-
-	PyObject *pyimg = PyTuple_GetItem(args, 1);
-	imgIn = PyString_AsString(pyimg);
-	Py_ssize_t imgLen = PyObject_Length(pyimg);
-
-	PyObject *pyPxFmt = PyTuple_GetItem(args, 2);
-	pxFmtIn = PyString_AsString(pyPxFmt);
-
-	PyObject *pyWidth = PyTuple_GetItem(args, 3);
-	widthIn = PyInt_AsLong(pyWidth);
-
-	PyObject *pyHeight = PyTuple_GetItem(args, 4);
-	heightIn = PyInt_AsLong(pyHeight);
-
-	std::map<std::string, class Video_out *>::iterator it = self->threads->find(devarg);
-
-	if(it != self->threads->end())
-	{
-		it->second->SendFrame(imgIn, imgLen, pxFmtIn, widthIn, heightIn);
-	}
-	else
-	{
-		PyErr_Format(PyExc_RuntimeError, "Device not found.");
-		Py_RETURN_NONE;
-	}
-
-	Py_RETURN_NONE;
-}
-
-PyObject *Video_out_manager_close(Video_out_manager *self, PyObject *args)
-{
-	//Process arguments
-	const char *devarg = "/dev/video0";
-	if(PyTuple_Size(args) >= 1)
-	{
-		PyObject *pydevarg = PyTuple_GetItem(args, 0);
-		devarg = PyString_AsString(pydevarg);
-	}
-
-	//Stop worker thread
-	std::map<std::string, class Video_out *>::iterator it = self->threads->find(devarg);
-
-	if(it != self->threads->end())
-	{
-		it->second->Stop();
-	}
-
-	Py_RETURN_NONE;
-}
-
-PyObject *Video_out_manager_list_devices(Video_out_manager *self)
-{
-	PyObject *out = PyList_New(0);
-	const char dir[] = "/dev";
-	DIR *dp;
-	struct dirent *dirp;
-	if((dp  = opendir(dir)) == NULL) {
-		printf("Error(%d) opening %s\n", errno, dir);
-		Py_RETURN_NONE;
-	}
-
-	while ((dirp = readdir(dp)) != NULL) {
-		if (strncmp(dirp->d_name, "video", 5) != 0) continue;
-		std::string tmp = "/dev/";
-		tmp.append(dirp->d_name);
-		PyList_Append(out, PyString_FromString(tmp.c_str()));
-	}
-	closedir(dp);
-
-	PyList_Sort(out);
-	return out;
-}
